0.6.0.1
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@39 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ mmex/include/stocks.h
@@ -83,7 +83,7 @@ public:
     double purchasePrice_;
     double value_;
     double gainLoss_;
-    long numShares_;
+    double numShares_;
 };
 
 class mmStocksPanel : public mmPanelBase
